+++ src/shellyDevice.test.ts
@@ -401,6 +401,87 @@ describe('Shelly devices test', () => {
     device?.destroy();
   });
 
+  test('create gen 2+ with wrong ws should log wrong settings', async () => {
+    shelly.ipv4Address = '192.168.1.20';
+    fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
+      if (service === 'shelly')
+        return Promise.resolve({
+          name: 'H&T Gen3',
+          id: 'shellyhtg3-3030f9ec8468',
+          mac: '3030F9EC8468',
+          model: 'S3SN-0U12A',
+          gen: 3,
+          fw_id: '20241011-121127/1.4.5-gbf870ca',
+          ver: '1.4.5',
+          auth_en: false,
+        });
+      if (service === 'Shelly.GetConfig')
+        return Promise.resolve({
+          ws: {
+            enable: true,
+            server: '192.168.1.XXX:8486',
+            ssl_ca: '*',
+          },
+          sys: {
+            device: {},
+            cfg_rev: 23,
+          },
+        });
+      if (service === 'Shelly.GetStatus')
+        return Promise.resolve({
+          sys: {
+            available_updates: {},
+            wakeup_period: 7200,
+          },
+        });
+      return Promise.resolve({ gen: 3 });
+    });
+    const host = '192.168.100.100';
+    const device = await ShellyDevice.create(shelly, log, host);
+    expect(device).toBeDefined();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.WARN, expect.stringContaining(`The Outbound websocket settings is not configured correctly`));
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.WARN, expect.stringContaining(`The port must be 8485`));
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.WARN, expect.stringContaining(`The ip must be the matterbridge ip`));
+    device?.destroy();
+  });
+
+  test('create gen 2+ without ws should log wrong settings', async () => {
+    shelly.ipv4Address = '192.168.1.20';
+    fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
+      if (service === 'shelly')
+        return Promise.resolve({
+          name: 'H&T Gen3',
+          id: 'shellyhtg3-3030f9ec8468',
+          mac: '3030F9EC8468',
+          model: 'S3SN-0U12A',
+          gen: 3,
+          fw_id: '20241011-121127/1.4.5-gbf870ca',
+          ver: '1.4.5',
+          auth_en: false,
+        });
+      if (service === 'Shelly.GetConfig')
+        return Promise.resolve({
+          sys: {
+            device: {},
+            cfg_rev: 23,
+          },
+        });
+      if (service === 'Shelly.GetStatus')
+        return Promise.resolve({
+          sys: {
+            available_updates: {},
+            wakeup_period: 7200,
+          },
+        });
+      return Promise.resolve({ gen: 3 });
+    });
+    const host = '192.168.100.100';
+    const device = await ShellyDevice.create(shelly, log, host);
+    expect(device).toBeDefined();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.stringContaining(`WebSocket server component not found`));
+    device?.destroy();
+  });
+
   test('create gen 1', async () => {
     fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
       if (service === 'shelly') return Promise.resolve({ type: 'SHSW-1', fw: '20210608-073743/v1.11.0@7b3d8b7d', mode: 'roller', auth: true });
@@ -418,6 +499,60 @@ describe('Shelly devices test', () => {
       LogLevel.NOTICE,
       expect.stringContaining(`Device ${hk}${device.id}${nt} host ${zb}${device.host}${nt} has an available firmware update.`),
     );
+
+    device.destroy();
+  });
+
+  test('create gen 1 with wrong coiot', async () => {
+    fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
+      if (service === 'shelly') return Promise.resolve({ type: 'SHSW-1', fw: '20210608-073743/v1.11.0@7b3d8b7d', mode: 'roller', auth: true });
+      if (service === 'status') return Promise.resolve({ has_update: true });
+      if (service === 'settings')
+        return Promise.resolve({ coiot: { enabled: false, peer: '192.168.100.101' }, device: { hostname: 'shellydevice-123456789' }, name: 'Shelly device' });
+      return Promise.resolve({});
+    });
+    const host = '192.168.100.100';
+    const device = await ShellyDevice.create(shelly, log, host);
+    expect(device).toBeDefined();
+    if (!device) return;
+    expect((device as any).lastseenInterval).toBeDefined();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.WARN, expect.stringContaining(`The CoIoT service is not enabled`));
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.WARN, expect.stringContaining(`is not mcast or`));
+
+    device.destroy();
+  });
+
+  test('create gen 1 with cover not calibrated', async () => {
+    fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
+      if (service === 'shelly') return Promise.resolve({ type: 'SHSW-1', fw: '20210608-073743/v1.11.0@7b3d8b7d', mode: 'roller', auth: true });
+      if (service === 'status') return Promise.resolve({ has_update: true });
+      if (service === 'settings') return Promise.resolve({ rollers: [{ current_pos: 101 }], device: { hostname: 'shellydevice-123456789' }, name: 'Shelly device' });
+      return Promise.resolve({});
+    });
+    const host = '192.168.100.100';
+    const device = await ShellyDevice.create(shelly, log, host);
+    expect(device).toBeDefined();
+    if (!device) return;
+    expect((device as any).lastseenInterval).toBeDefined();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.NOTICE, expect.stringContaining(`does not have position control enabled`));
+
+    device.destroy();
+  });
+
+  test('create gen 2 with cover not calibrated', async () => {
+    fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
+      if (service === 'shelly')
+        return Promise.resolve({ type: 'SHSW-1', id: 'shellyhtg3-3030f9ec8468', fw_id: '20210608-073743/v1.11.0@7b3d8b7d', mode: 'roller', auth: true, gen: 2 });
+      if (service === 'Shelly.GetStatus') return Promise.resolve({ has_update: false, sys: { device: {}, cfg_rev: 23, available_updates: {} } });
+      if (service === 'Shelly.GetConfig') return Promise.resolve({ 'cover:0': { pos_control: false }, 'device': { hostname: 'shellydevice-123456789' }, 'name': 'Shelly device' });
+      return Promise.resolve({});
+    });
+    const host = '192.168.100.100';
+    const device = await ShellyDevice.create(shelly, log, host);
+    expect(device).toBeDefined();
+    if (!device) return;
+    expect((device as any).lastseenInterval).toBeDefined();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.NOTICE, expect.stringContaining(`does not have position control enabled`));
 
     device.destroy();
   });
@@ -715,6 +850,23 @@ describe('Shelly devices test', () => {
     device.destroy();
   });
 
+  test('onUpdate should handle bthomesensor update with unknown sensor', async () => {
+    const device = await ShellyDevice.create(shelly, log, path.join('src', 'mock', 'shelly2pmg3-34CDB0770C4C.json'));
+    if (!device) return;
+
+    const updateData = {
+      ['bthomesensor:invalid']: {
+        value: 25.5,
+        last_updated_ts: 1625072400,
+      },
+    };
+
+    device.onUpdate(updateData);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, expect.stringContaining(`*Unknown bthomesensor`));
+
+    device.destroy();
+  });
+
   test('onUpdate should handle gen 1 device data', async () => {
     fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
       if (service === 'shelly') return Promise.resolve({ type: 'SHSW-1', fw: '20210608-073743/v1.11.0@7b3d8b7d', auth: false });
@@ -779,6 +931,39 @@ describe('Shelly devices test', () => {
 
     const result = await device.saveDevicePayloads('temp');
     expect(result).toBe(true);
+
+    device.destroy();
+  });
+
+  test('saveDevicePayloads should handle errors', async () => {
+    fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
+      if (service === 'shelly') return Promise.resolve({ type: 'SHSW-1', fw: '20210608-073743/v1.11.0@7b3d8b7d', auth: false });
+      if (service === 'status')
+        return Promise.resolve({
+          has_update: false,
+          wifi_ap: { ssid: 'secret_ap' },
+          wifi_sta: { ssid: 'secret_sta' },
+          wifi_sta1: { ssid: 'secret_sta1' },
+        });
+      if (service === 'settings')
+        return Promise.resolve({
+          device: { hostname: 'shellydevice-123456789' },
+          name: 'Shelly device',
+          timezone: 'Europe/Rome',
+          lat: 45.123,
+          lng: 9.456,
+          wifi_ap: { ssid: 'secret_ap_settings' },
+          wifi_sta: { ssid: 'secret_sta_settings' },
+          wifi_sta1: { ssid: 'secret_sta1_settings' },
+        });
+      return Promise.resolve({});
+    });
+
+    const device = await ShellyDevice.create(shelly, log, '192.168.1.100');
+    if (!device) return;
+
+    const result = await device.saveDevicePayloads('temp<><<++"""');
+    expect(result).toBe(false);
 
     device.destroy();
   });